Lite provider
Writing Content Providers in Android Applications can be really
annoying. In most cases there is not much complexity in regards of the
data. Still, you have to write a great deal of boilerplate code to get
going. This library is intended to ease up the creation of very simple
Content Providers using annotations.

Signed-off-by: Solo <anqi.huang@outlook.com>